refactor: divide calcOffsetModification
